+++ CHANGELOG.md
@@ -3,10 +3,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.7.1] - 2025-06-01
+- Math conversion now includes the MVR (move remainder) opcode.
+- Better handling of blank comment (e.g., "C*") and blank lines embedded between specs.
+- Some math opcodes with resulting indicator use (which is very rare) were returning Resulting Indicator 1 in all cases. This has been corrected.
 ## [0.6.43] - 2025-05-31
 - Corrected resulting indicator conversion for database I/O opcodes
 - Added `TESTZ`, `TESTB`, `TESTN`,`BITON`, and `BITOFF` opcodes to the conversion.
-- Corrected an issue with Formatted output related to hexadecimal and binary notation.
+- Corrected an issue with conversion output related to hexadecimal and binary notation literals.
 ## [0.6.42] - 2025-05-31
 - Enhanced the SUBST (substring) opcode conversion routine to match our contemporary design style. In some cases the resulting %SUBST will be more streamlined, but the result will be the same.
 ## [0.6.40] - 2025-05-31